[fix]: upload consulting and deleting

+++ src/modules/proposals/useCases/deleteUpload/deleteUploadUseCase.ts
@@ -1,3 +1,4 @@
+import { IStorageProvider } from '@shared/container/providers/StorageProvider/IStorageProvider';
 import { IProposalUploadsRepository } from "@modules/proposals/repositories/IProposalUploadsRepository";
 import { inject, injectable } from "tsyringe";
 
@@ -6,10 +7,20 @@ import { inject, injectable } from "tsyringe";
 class DeleteUploadUseCase {
   constructor(
     @inject("ProposalUploadsRepository")
-    private proposalUploadsRepository: IProposalUploadsRepository
+    private proposalsUploadsRepository: IProposalUploadsRepository,
+    @inject("ProposalUploadsRepository")
+    private proposalUploadsRepository: IProposalUploadsRepository,
+    @inject("StorageProvider")
+    private storageProvider: IStorageProvider
   ) {}
   async execute(id: number, upload_id: string): Promise<void> {
-    await this.proposalUploadsRepository.delete(id, upload_id);
+    let upload = await this.proposalsUploadsRepository.findById(id, upload_id);
+
+    if(upload) {
+      await this.proposalUploadsRepository.delete(id, upload_id);
+
+      await this.storageProvider.delete(upload.upload_name, 'uploads')
+    }
   }
 }
